backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
numpy
torch
torchvision
matplotlib
opencv-python
